# common-library/src/main/java/cz/fi/muni/pa165/dto/userService/PaymentBudgetPackageOfferDto.java
package cz.fi.muni.pa165.dto.userService;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.UUID;

@Setter
@Getter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class PaymentBudgetPackageOfferDto {

	private UUID guid;

	private Integer priceDollars;

	private Integer budgetIncrease;

	private String description;

	private Boolean isAvailable;

}