# openapi/user-service.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8083/api
  description: Generated server url
...
paths:
...
  /v1/role/:
    put:
      tags:
      - Role API
      description: Updates role by it's ID
      operationId: updateRole
      requestBody:
        description: Updated role values
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleViewDto"
        required: true
      responses:
        "400":
          description: Request body does not meet validation requirements
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired role does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: "Role was successfully updated, returning updated model"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleViewDto"
    post:
      tags:
      - Role API
      description: Creates a new role with given properties
      operationId: createRole
      requestBody:
        description: Data for role creation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleViewDto"
        required: true
      responses:
        "400":
          description: Validation of input data failed
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Entity already exists
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "201":
          description: Role View with newly created role data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleViewDto"
  /v1/payment/:
    put:
      tags:
      - Payment API
      description: Updates payment by it's ID
      operationId: updatePayment
      requestBody:
        description: Updated payment values
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdateCreateDto"
        required: true
      responses:
        "400":
          description: Request body does not meet validation requirements
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired payment does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: "Payment was successfully updated, returning updated model"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentViewDto"
    post:
      tags:
      - Payment API
      description: Creates a new Payment with given properties
      operationId: createPayment
      requestBody:
        description: Data for payment creation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdateCreateDto"
        required: true
      responses:
        "400":
          description: Validation of input data failed
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Entity already exists
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired user or budget package does not exist
          content:
            application/json:
              schema:
                type: string
        "201":
          description: User View with newly created payment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentViewDto"
...
  /v1/role/{id}:
    get:
      tags:
      - Role API
      description: Returns desired role by its id.
      operationId: getRoleById
      parameters:
      - name: id
        in: path
        description: UUID of desired role
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Path parameter does not have a form of UUID
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired role does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Role view with desired details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleViewDto"
    delete:
      tags:
      - Role API
      description: Deletes role by it's ID
      operationId: deleteRole
      parameters:
      - name: id
        in: path
        description: Id of role that should be deleted.
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired role does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Role was successfully deleted.
          content:
            application/json:
              schema:
                type: boolean
  /v1/role/all-roles:
    get:
      tags:
      - Role API
      description: Returns all available roles.
      operationId: getAllRoles
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: List of all roles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleViewDto"
  /v1/payment/{id}:
    get:
      tags:
      - Payment API
      description: Returns desired payment by its id.
      operationId: getPaymentById
      parameters:
      - name: id
        in: path
        description: UUID of desired payment
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Path parameter does not have a form of UUID
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired payment does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Payment view with desired details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentViewDto"
  /v1/payment/all-payments:
    get:
      tags:
      - Payment API
      description: Returns all payments.
      operationId: getAllPayments
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: List of all payments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentViewDto"
...
components:
  schemas:
    PaymentBudgetPackageOfferDto:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        priceDollars:
          type: integer
          format: int32
        budgetIncrease:
          type: integer
          format: int32
        description:
          type: string
        isAvailable:
          type: boolean
    RoleViewDto:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        code:
          type: string
          maxLength: 50
          minLength: 0
      required:
      - code
      - description
      - name
...
    PaymentUpdateCreateDto:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        budgetOfferPackageId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        paid:
          type: boolean
    PaymentUserViewDto:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        username:
          type: string
        mail:
          type: string
        isActive:
          type: boolean
        name:
          type: string
        surname:
          type: string
    PaymentViewDto:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        user:
          $ref: "#/components/schemas/PaymentUserViewDto"
        budgetOfferPackage:
          $ref: "#/components/schemas/PaymentBudgetPackageOfferDto"
        createdAt:
          type: string
          format: date-time
        paid:
          type: boolean
...